mnd_lms_simple/requirements.txt
# MND LMS Requirements

Django==5.2.8
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
Pillow==10.4.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# Development dependencies
django-debug-toolbar==4.2.0
coverage==7.3.2
black==23.9.1
flake8==6.1.0

# Production dependencies for Render
dj-database-url==2.1.0
python-dotenv==1.0.0